refactor: merge AuthorityBase into Authority
- Delete AuthorityBase.java (@MappedSuperclass); inline all fields,
  constants, and methods directly into Authority.java
- Authority now extends MetadataEntity directly
- Remove AuthorityBase.class::cast in delegate mappers; Page<Authority>
  is passed directly to the mapper
- Update AuthorityRowMapper and AuthorityReindexJobRunner static imports
  from AuthorityBase.X_COLUMN to Authority.X_COLUMN
- Add missing Authority import in AuthorityUtilityMapper
- Remove unused Authority import from AuthorityCqlRepository
- Fix extra blank line in AuthorityMapperTest import block
