+++ src/main/java/org/folio/entlinks/domain/repository/AuthorityBaseCqlRepositoryImpl.java
@@ -3,31 +3,30 @@ package org.folio.entlinks.domain.repository;
 import jakarta.persistence.EntityManager;
 import java.util.List;
 import java.util.UUID;
-import org.folio.entlinks.domain.entity.AuthorityBase;
+import org.folio.entlinks.domain.entity.Authority;
 import org.folio.spring.cql.Cql2JpaCriteria;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.support.PageableExecutionUtils;
 
-public abstract class AuthorityBaseCqlRepositoryImpl<T extends AuthorityBase> implements AuthorityBaseCqlRepository<T> {
+public class AuthorityBaseCqlRepositoryImpl implements AuthorityBaseCqlRepository {
 
   private final EntityManager em;
-  private final Cql2JpaCriteria<T> cql2JpaCriteria;
+  private final Cql2JpaCriteria<Authority> cql2JpaCriteria;
 
-  protected AuthorityBaseCqlRepositoryImpl(EntityManager em) {
+  public AuthorityBaseCqlRepositoryImpl(EntityManager em) {
     this.em = em;
-    this.cql2JpaCriteria = new Cql2JpaCriteria<>(getClassType(), em);
+    this.cql2JpaCriteria = new Cql2JpaCriteria<>(Authority.class, em);
   }
 
   @Override
-  public Page<T> findByCql(String cqlQuery, Pageable pageable) {
-
-    var collectBy = collectByQuery(cqlQuery);
-    var countBy = countByQuery(cqlQuery);
+  public Page<Authority> findByCql(String cqlQuery, Pageable pageable) {
+    var collectBy = deletedIs(Boolean.FALSE).and(cql2JpaCriteria.createCollectSpecification(cqlQuery));
+    var countBy = deletedIs(Boolean.FALSE).and(cql2JpaCriteria.createCountSpecification(cqlQuery));
     var criteria = cql2JpaCriteria.toCollectCriteria(collectBy);
 
-    List<T> resultList = em
+    List<Authority> resultList = em
       .createQuery(criteria)
       .setFirstResult((int) pageable.getOffset())
       .setMaxResults(pageable.getPageSize())
@@ -37,14 +36,14 @@ public abstract class AuthorityBaseCqlRepositoryImpl<T extends AuthorityBase> im
 
   @Override
   public Page<UUID> findIdsByCql(String cqlQuery, Pageable pageable) {
-    var collectBy = collectByQuery(cqlQuery);
-    var countBy = countByQuery(cqlQuery);
+    var collectBy = deletedIs(Boolean.FALSE).and(cql2JpaCriteria.createCollectSpecification(cqlQuery));
+    var countBy = deletedIs(Boolean.FALSE).and(cql2JpaCriteria.createCountSpecification(cqlQuery));
 
     var cb = em.getCriteriaBuilder();
     var query = cb.createQuery(UUID.class);
-    var root = query.from(getClassType());
+    var root = query.from(Authority.class);
 
-    query.select(root.get(AuthorityBase.ID_COLUMN));
+    query.select(root.get(Authority.ID_COLUMN));
     query.where(collectBy.toPredicate(root, query, cb));
 
     List<UUID> resultList = em
@@ -56,12 +55,12 @@ public abstract class AuthorityBaseCqlRepositoryImpl<T extends AuthorityBase> im
   }
 
   @Override
-  public Page<T> findDeletedByCql(String cqlQuery, Pageable pageable) {
+  public Page<Authority> findDeletedByCql(String cqlQuery, Pageable pageable) {
     var collectBy = deletedIs(Boolean.TRUE).and(cql2JpaCriteria.createCollectSpecification(cqlQuery));
     var countBy = deletedIs(Boolean.TRUE).and(cql2JpaCriteria.createCountSpecification(cqlQuery));
     var criteria = cql2JpaCriteria.toCollectCriteria(collectBy);
 
-    List<T> resultList = em
+    List<Authority> resultList = em
       .createQuery(criteria)
       .setFirstResult((int) pageable.getOffset())
       .setMaxResults(pageable.getPageSize())
@@ -76,9 +75,9 @@ public abstract class AuthorityBaseCqlRepositoryImpl<T extends AuthorityBase> im
 
     var cb = em.getCriteriaBuilder();
     var query = cb.createQuery(UUID.class);
-    var root = query.from(getClassType());
+    var root = query.from(Authority.class);
 
-    query.select(root.get(AuthorityBase.ID_COLUMN));
+    query.select(root.get(Authority.ID_COLUMN));
     query.where(collectBy.toPredicate(root, query, cb));
 
     List<UUID> resultList = em
@@ -89,24 +88,12 @@ public abstract class AuthorityBaseCqlRepositoryImpl<T extends AuthorityBase> im
     return PageableExecutionUtils.getPage(resultList, pageable, () -> count(countBy));
   }
 
-  protected abstract Class<T> getClassType();
-
-  protected abstract Boolean deleted();
-
-  private long count(Specification<T> specification) {
+  private long count(Specification<Authority> specification) {
     var criteria = cql2JpaCriteria.toCountCriteria(specification);
     return em.createQuery(criteria).getSingleResult();
   }
 
-  private Specification<T> collectByQuery(String cqlQuery) {
-    return deletedIs(deleted()).and(cql2JpaCriteria.createCollectSpecification(cqlQuery));
-  }
-
-  private Specification<T> countByQuery(String cqlQuery) {
-    return deletedIs(deleted()).and(cql2JpaCriteria.createCountSpecification(cqlQuery));
-  }
-
-  private Specification<T> deletedIs(Boolean deleted) {
-    return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get(AuthorityBase.DELETED_COLUMN), deleted);
+  private Specification<Authority> deletedIs(Boolean deleted) {
+    return (root, query, criteriaBuilder) -> criteriaBuilder.equal(root.get(Authority.DELETED_COLUMN), deleted);
   }
 }
