+++ src/main/java/org/folio/entlinks/domain/repository/AuthorityCqlRepository.java
@@ -1,7 +1,5 @@
 package org.folio.entlinks.domain.repository;
 
-import org.folio.entlinks.domain.entity.Authority;
-
-public interface AuthorityCqlRepository extends AuthorityBaseCqlRepository<Authority> {
+public interface AuthorityCqlRepository extends AuthorityBaseCqlRepository {
 
 }
